+++ lib/rubocop/cop/rails/ordered_macros.rb
@@ -28,7 +28,6 @@ module RuboCop
       #
       class OrderedMacros < ::RuboCop::Cop::Base
         MSG = 'Sort `%<macro>s` declarations alphabetically (`%<name>s` should come before `%<previous>s`).'
-
         MACROS = %i[belongs_to has_one has_many has_and_belongs_to_many validates scope].freeze
 
         def self.default_configuration