+++ lib/rubocop/cop/rspec/overwritten_let.rb
@@ -33,7 +33,6 @@ module RuboCop
       #
       class OverwrittenLet < ::RuboCop::Cop::Base
         MSG = 'Do not override the outer `let` `%<name>s`; use a distinct name or set the value in `before`.'
-
         LET_METHODS = %i[let let!].freeze
         EXAMPLE_GROUP_METHODS = %i[describe context].freeze